简单的实现
